Reverted all market changes. ---buyorder update/logging ---sellorder update/logging ---markterorder update/logging

+++ src/eve-server/character/Character.h
@@ -220,7 +220,7 @@ public:
     /**
      * Primary public interface:
      */
-    bool            AlterBalance(double amount, uint8 type);  // ---sellorder update
+    bool            AlterBalance(float amount, uint8 type);
     void            SetLocation(uint32 stationID, SystemData& data);
     void            JoinCorporation(const CorpData& data);
     void            SetDescription(const char *newDescription);
